mm: refactor to support paging

+++ lib/stdio_impl.c
@@ -518,7 +518,7 @@ int vsnprintf(char *buf, size_t size, const char *format, va_list args)
                 case 'p':
                 {
                     // print a pointer
-                    u64 value = (u64) va_arg(args, void *);
+                    u64 value = (size_t) va_arg(args, void *);
                     buf_putchar(&buf, '0');
                     buf_putchar(&buf, 'x');
                     int c = printf_diouxX(buf, value, &flags, 'x');
